+++ sdk/src/math/margin.ts
@@ -166,10 +166,10 @@ export function calculateWorstCasePerpLiabilityValue(
 	perpPosition: PerpPosition,
 	perpMarket: PerpMarketAccount,
 	oraclePrice: BN,
-	includeOpenOrders = true
+	includeOpenOrders: boolean = true
 ): { worstCaseBaseAssetAmount: BN; worstCaseLiabilityValue: BN } {
 	const isPredictionMarket = isVariant(perpMarket.contractType, 'prediction');
-
+	// return early if no open orders required
 	if (!includeOpenOrders) {
 		return {
 			worstCaseBaseAssetAmount: perpPosition.baseAssetAmount,
@@ -180,7 +180,6 @@ export function calculateWorstCasePerpLiabilityValue(
 			),
 		};
 	}
-
 	const allBids = perpPosition.baseAssetAmount.add(perpPosition.openBids);
 	const allAsks = perpPosition.baseAssetAmount.add(perpPosition.openAsks);