+++ sdk/src/pyth/constants.ts
@@ -1,0 +1,9 @@
+import { PublicKey } from '@solana/web3.js';
+
+export const DEFAULT_WORMHOLE_PROGRAM_ID = new PublicKey(
+	'HDwcJBJXjL9FpJ7UBsYBtaDjsBUhuLCUYoz3zr8SWWaQ'
+);
+
+export const DEFAULT_RECEIVER_PROGRAM_ID = new PublicKey(
+	'rec5EKMGg6MxZYaMdyBfgwp4d5rB9T1VQH5pJv5LtFJ'
+);